Add sentry support
