+++ lib/utils/queues.js
@@ -1,4 +1,5 @@
 const Queue = require('bull')
+const sentry = require('./sentry')
 
 const checkQueue = new Queue('check', {
   redis: {
@@ -9,7 +10,7 @@ const checkQueue = new Queue('check', {
 })
 
 checkQueue.on('failed', (job, err) => {
-  console.log(err)
+  sentry.captureException(err)
 })
 
 module.exports = {checkQueue}
