+++ worker.js
@@ -1,10 +1,16 @@
+const sentry = require('./lib/utils/sentry')
 const mongo = require('./lib/utils/mongo')
+const {checkQueue} = require('./lib/utils/queues')
+
+const check = require('./jobs/check')
 
 async function main() {
+  await checkQueue.isReady()
   await mongo.connect()
 
-  // eslint-disable-next-line import/no-unassigned-import
-  require('./jobs/check')
+  checkQueue.process(({data: {linkId, name: location}}) => check(linkId, location))
 }
 
-main()
+main().catch(err => {
+  sentry.captureException(err)
+})
